refactor: 3단계 피드백 반영
