+++ src/test/java/racinggame/domain/NeverMoveCondition.java
@@ -1,13 +1,9 @@
 package racinggame.domain;
 
-import java.util.Random;
-
 public class NeverMoveCondition extends MoveCondition {
 
-    private static final int LEAST_MOVABLE_CONDITION = 4;
-    private static final Random RANDOM = new Random();
-
-    public NeverMoveCondition() {
-        super(RANDOM.nextInt(LEAST_MOVABLE_CONDITION));
+    @Override
+    public boolean isMovable() {
+        return false;
     }
 }
